moved piece of code in separate component
